Move the pattern rules into patterns.py, with section 4.2
