+++ src/patterns.py
@@ -3,7 +3,11 @@ from __future__ import annotations
 import ast
 from typing import Optional
 
-from contexts import ModuleContext, ancestors, class_entry, field_map, fields
+import reasons
+from reasons import IllFormedModule
+from contexts import (ModuleContext, ancestors, class_entry, field_map, fields,
+                      short_name)
+from aux import binds_seq, qualified_name
 
 def is_catch_all(p: ast.pattern) -> bool:
     return isinstance(p, ast.MatchAs) and p.pattern is None
@@ -57,3 +61,45 @@ def subsumes(p: ast.pattern, q: ast.pattern, ctx: ModuleContext) -> bool:
             return False
         return all(subsumes(map_p[x], map_q[x], ctx) for x in fields(c_q))
     return False
+
+def check_pattern(p: ast.pattern, ctx: ModuleContext) -> None:
+    if isinstance(p, ast.MatchClass):
+        entry = class_entry(p.cls, ctx)
+        if entry is None:
+            raise IllFormedModule(p, reasons.UnknownClassInPattern(qualified_name(p.cls) if isinstance(p.cls, (ast.Name, ast.Attribute)) else ast.unparse(p.cls)))
+        c_name, xs = short_name(entry), fields(entry)
+        if field_map(entry, p.patterns, p.kwd_attrs, p.kwd_patterns) is None:
+            n = len(p.patterns)
+            if n + len(p.kwd_attrs) != len(xs):
+                raise IllFormedModule(p, reasons.PatternArityMismatch(c_name, len(xs), n + len(p.kwd_attrs)))
+            if len(p.kwd_attrs) != len(set(p.kwd_attrs)):
+                raise IllFormedModule(p, reasons.DuplicatePatternKeyword(c_name))
+            raise IllFormedModule(p, reasons.UnknownFieldInPattern(c_name, tuple(sorted(set(xs[n:])))))
+        for sub in list(p.patterns) + list(p.kwd_patterns):
+            check_pattern(sub, ctx)
+        return
+    if isinstance(p, ast.MatchSequence):
+        for sub in p.patterns:
+            check_pattern(sub, ctx)
+        return
+    if isinstance(p, ast.MatchMapping):
+        keys = [dict_key(key) for key in p.keys]
+        duplicate = next((k for i, k in enumerate(keys) if k in keys[:i]), None)
+        if duplicate is not None:
+            raise IllFormedModule(p, reasons.DuplicateDictKey(duplicate))
+        for sub in p.patterns:
+            check_pattern(sub, ctx)
+        return
+    if isinstance(p, ast.MatchAs) and p.pattern is not None:
+        check_pattern(p.pattern, ctx)
+        return
+
+def check_pattern_list(patterns: list[ast.pattern], node: ast.AST, ctx: ModuleContext) -> None:
+    for i, p in enumerate(patterns):
+        check_pattern(p, ctx)
+        vars_ = binds_seq(p)
+        if len(vars_) != len(set(vars_)):
+            raise IllFormedModule(node, reasons.NonlinearPattern(i + 1))
+        for j in range(i):
+            if subsumes(p, patterns[j], ctx):
+                raise IllFormedModule(node, reasons.UnreachableCase(i + 1, j + 1))
