+++ src/well_formed.py
@@ -7,12 +7,13 @@ import reasons
 from reasons import IllFormedModule
 from contexts import (ClassEntry, Context, ContextEntry, ModuleContext, ModuleLoaded, ModuleStub,
                       class_entry, entry_of,
-                      field_map, ResultTy, Status, ASSIGNS_EMPTY, RETURNS, Assigns,
+                      field_map, short_name, ResultTy, Status, ASSIGNS_EMPTY, RETURNS, Assigns,
                       Returns, VarContext, class_of, extend_context, fields, merge_results,
                       module_of, override_gamma, override_results, override_var, var_status)
 from aux import (BlockElement, assigns_block, assigns_elements, binds, captures_e, captures_element,
-                 binds_seq, elements_of_block, find_first_reassigning, names_in_target, own_fields)
-from patterns import dict_key, is_catch_all, subsumes
+                 elements_of_block, find_first_reassigning, names_in_target, own_fields,
+                 qualified_name)
+from patterns import check_pattern_list, dict_key, is_catch_all
 
 def class_entry_for(node: ast.ClassDef, q: str, context: Context) -> ClassEntry:
     base = node.bases[0].id if node.bases and isinstance(node.bases[0], ast.Name) else None
@@ -263,57 +264,6 @@ def check_exprs(es: list[ast.expr], ctx: ModuleContext) -> None:
         return
     check_expr(es[0], ctx)
     check_exprs(es[1:], ctx)
-
-def short_name(entry: ClassEntry) -> str:
-    return entry.name.rsplit('.', 1)[-1]
-
-def qualified_name(e: ast.expr) -> str:
-    if isinstance(e, ast.Name):
-        return e.id
-    assert isinstance(e, ast.Attribute)
-    return qualified_name(e.value) + '.' + e.attr
-
-def check_pattern(p: ast.pattern, ctx: ModuleContext) -> None:
-    if isinstance(p, ast.MatchClass):
-        entry = class_entry(p.cls, ctx)
-        if entry is None:
-            raise IllFormedModule(p, reasons.UnknownClassInPattern(qualified_name(p.cls) if isinstance(p.cls, (ast.Name, ast.Attribute)) else ast.unparse(p.cls)))
-        c_name, xs = short_name(entry), fields(entry)
-        if field_map(entry, p.patterns, p.kwd_attrs, p.kwd_patterns) is None:
-            n = len(p.patterns)
-            if n + len(p.kwd_attrs) != len(xs):
-                raise IllFormedModule(p, reasons.PatternArityMismatch(c_name, len(xs), n + len(p.kwd_attrs)))
-            if len(p.kwd_attrs) != len(set(p.kwd_attrs)):
-                raise IllFormedModule(p, reasons.DuplicatePatternKeyword(c_name))
-            raise IllFormedModule(p, reasons.UnknownFieldInPattern(c_name, tuple(sorted(set(xs[n:])))))
-        for sub in list(p.patterns) + list(p.kwd_patterns):
-            check_pattern(sub, ctx)
-        return
-    if isinstance(p, ast.MatchSequence):
-        for sub in p.patterns:
-            check_pattern(sub, ctx)
-        return
-    if isinstance(p, ast.MatchMapping):
-        keys = [dict_key(key) for key in p.keys]
-        duplicate = next((k for i, k in enumerate(keys) if k in keys[:i]), None)
-        if duplicate is not None:
-            raise IllFormedModule(p, reasons.DuplicateDictKey(duplicate))
-        for sub in p.patterns:
-            check_pattern(sub, ctx)
-        return
-    if isinstance(p, ast.MatchAs) and p.pattern is not None:
-        check_pattern(p.pattern, ctx)
-        return
-
-def check_pattern_list(patterns: list[ast.pattern], node: ast.AST, ctx: ModuleContext) -> None:
-    for i, p in enumerate(patterns):
-        check_pattern(p, ctx)
-        vars_ = binds_seq(p)
-        if len(vars_) != len(set(vars_)):
-            raise IllFormedModule(node, reasons.NonlinearPattern(i + 1))
-        for j in range(i):
-            if subsumes(p, patterns[j], ctx):
-                raise IllFormedModule(node, reasons.UnreachableCase(i + 1, j + 1))
 
 def check_class_decl(node: ast.ClassDef, gamma: Context, q: str) -> None:
     if isinstance(gamma.get(node.name), ClassEntry):
